Get OAuth information from Provider rather than Domain
Previously the OAuth provider was chosen based on the
account's domain. This would fail for custom domain
names hosted under services such as Google Apps for
Business or the like.

Here we take a Provider and look it up in the config
to get the OAuth url, but fall back to domain if none
is found for backwards compatibility.